NotifyWatcher -> StringsWatcher
The following code changes the LXD Profile watcher from a Notify
to a Strings watcher. In the current code the notify watcher fires
the event, and then it responds with another request to fetch the
upgrade status. This was legacy code before we move to the uniter,
as the command for upgrade charm didn't work with a StringsWatcher.

Now that the command code is no longer used, we can utilise the
StringsWatcher and send the status with every trigger, reducing
latency and possible race conditions.