Place missing upgrade charm profile data
The following test expects that the instance charm profile data
should exist, so that the upgrade charm profile resolver can process.
Unfortunately the new changes made it so NotKnowStatus is no longer
terminal, which then means that all upgrades NEED to have a document.

We know this is the case from the code point of view, but the
test setup didn't know this, so we have to add code to make sure
that it does.

This is generalised enough to be used in other tests if needs be.