Un-place missing upgrade charm profile data
This is a revert of 38b5255759be8578010c04a132cb23224f8c2789 with
added notes to prevent people from making the same mistake that I
did.

Although the code is generalised for other users to use, if nobody
else is using it, then we should delete it for the sake of our
sanity.