Switch a lot of watchers that don't care about revno.
Many of the state/watcher.go implementations were only reading the
document so they could call Watch(). However, as they aren't doing
anything with the initial document, they can't really care about the
revno. (changes since I looked at the db are meaningless if you aren't
tracking what the content was when you read it.)

The only caveat is that they are at least seeing that the document
*exists*. However, the first change that comes through should be
'revno: -1', which should tell them that the document has been deleted.
Which should be identical to starting the watcher just before it is
deleted, and then it is deleted before the loop engages.