Deploy a default IAAS bundle
Instead of deploying a kubernetes core for all stacks, instead
pick a different default bundle per stack. So for IAAS, the stack
is based on https://github.com/conjure-up/spells/blob/master/openstack-novalxd/bundle.yaml
with changes to ensure that it has a better chance of deploying
on the CI setup.