Individual logsink metric gauge.
The following creates the basis for adding metrics to the logsink
API. This is partially ground work to then add future changes to
how we construct the api server metrics.

Note I've also removed the connection pause time metric, as that
was harded coded to always return 0 and the commented code was just
left there in place. I can reinstate it, but we should have a better
backing for that metric instead of 0.