Remove prometheus register
The following commit removes the prometheus register from the api
server package. Instead we instantiate the metrics at the worker
level and then thread it through the configurations to the api
server.

The cleans up a lot of the code around how we instantiate the
metrics, by removing the out of order code of Unregister/Register.
It will always attempt to Register then Unregister, with the worker
correctly cleanup the metric collector from the prometheus
registry.

The following also includes TODO's for removing the Deprecated
metric collectors for release 2.6