Emit fixup ops in raftlease trapdoor if assertion fails
If there's an error writing a claim to the database there isn't any way
of retrying in notifyTarget. Instead, if the trapdoor assertion fails on
the first attempt, we check whether the leaseholders collection is
in sync, and if not we write out the same ops we'd normally have used to
claim the lease. This is safe to do because the trapdoor is only used
from places that are already using token.Check, which would fail if raft
didn't think the requested holder held the lease. But it is a bit dirty.