Get rid of HubWatcher.current revno cache.
Now that our state/watcher.go implementations expect that WatchNoRevno()
won't generate an event unless there is a real event, we can get rid of
the cache. All the callers have been audited for the new behavior.
We also use collect() in the initial event for docWatcher because that
allows better behavior around 'there is an event about to trigger' and
we also create our automatic first event.

Propogate the behavior of "Watch will sync before returning" to the
Watcher/BaseWatcher behavior as well.