Cleanup the BaseWatcher interface.
Document what is different at this layer, and get rid of WatchAtRevno
and rename WatchNoRevno back to just Watch().
The biggest difference is around initial event vs synchronizing that
Watch() won't return until the watcher is installed. This way watchers
can trust the Revno they get in the watch events is correct, without
having to cache the revno ahead of time.