Ensure we use the right modelUUID
The previous code wrongly selected the model UUID from the API
server. Except that modelUUID wasn't actually being instantiated
at all, so was probably old code that had been refactored, but
wasn't deleted. Instead now we use the approach found in the
api server, where by we interrogate the http request for the model
UUID. If the model UUID isn't found, we fall back to the controller
UUID and then use that in the metrics.