IsExposed() is a network round trip.
There is nothing about application.Refresh() that caches the exposed
information about the application. As such, calling IsExposed is an
independent api call, and can just as easily fail because the
application has gone away in the mean time. application.Refresh() only
refreshes the life attribute, and we don't even check the
application.Life() after calling Refresh, I'm not sure that we should
keep the previous call.