Ensure we still get raft metrics after a bounce
The go-metrics library registers the returned collector with the default
registry (which we don't use). If we don't unregister it then the
we lose metrics after the worker restarts, for example if a mongo i/o
timeout causes the state worker to bounce.

When raft times out starting up we can get a situation where a second
loop is started before the old one finishes (so it hasn't yet
unregistered). Use unregister/register rather than
register/defer-unregister to handle this.