Found an actual bug.
The issue is that if a unit assigned to a machine was made dead and
removed before we saw the event, we would try to read the doc and not
find it, and then try to remove the unit named "" rather than the unit
that we saw the event for.

Also fix a bug in StringsWatcherC if you are expecting multiple strings
and you get only 1 of them.

Fix an intermittent failure in another test.