Omit model list from show-controller output when user lacks permission
After going through the controller code that deals with permission
granting we noticed that the 'add-model' permission gets handled
differently (as a cloud controller permission instead of a controller
permission).

While the user may be granted 'add-model', the user access code in
showcontroller returns back a 'login' permission even though the user
should be technically able to list the models they have access to. In
this scenario the solution introduced by the previous commit would
incorrectly return an error message to the user.

As the controller information may be locally cached we thought there is
no harm in showing them to the user so the fix from the previous commit
has been updated to default to an empty model list if the user lacks the
required permissions; users with 'add-model' permissions can use the
list-models command to query the models available to them.