Actually fail properly under more cases.
The existing infrastructure was trying to fail after 10s if the test was
not making all of the Store calls that it wanted. However, the timer
didn't start until the test returned, which is weird. This new path
starts a stronger timer that will kill both the Store and the Manager if
the test takes >10s. I know this works because I have tests that needed
fixing :).