Wait on retries for all types of reasons.
The existing code would only wait to retry an API call if it got
Timeout. However, we can also get a ErrInvalid, which we would then
retry in a tight loop. Now we reuse the retry loop to slow down the
second and subsequent attempts to claim.

This lead to some testing hairyness, as the attempt loop was now tied to
the testing clock, that doesn't advance unless you find the right times
to do so. It also uncovered ways to cause the test to deadlock, and not
properly timeout after 10s.