Get some tests passing.
This needs:
  https://github.com/juju/clock/pull/4

In order to pass without hanging. This is because the
ExpiryRepeatedTimeout test sets up a situation where there is a lease
that has already expired by the time we enter loop(). The
testclock.Timer was properly being set to <=0 seconds, but Reset()
wasn't triggering the timer without calling WaitAdvance(0,0,1).