Get most of the tests passing.
ExpiryTimeout is now failing. Need to figure out if that should be a
source code change or a test suite change.

We wanted to give a bit of time for the expiry to run so that our Leases
would be fresh before we evalute them for blocks and the next expire
evaluation time. However, that means that a blocking Expire call will
prevent us from running the main loop.