We can potentially add a timer for how long we're willing to wait.
If we use NewTimer() we have the ability to Stop() the timer, which will
remove it from the wait list, which means it doesn't make it
unpredictable how many timers we have. Either we are testing waiting on
that timer, or the other cause triggers and stops the timer, so it
doesn't show up as a waiting timer.

Breaks other tests, though. We may just want to not synchronize with
retryingExpire.