Don't apply when not provisioned
The following changes ensure that we don't attempt to apply a
charm profile to a machine that isn't either running or that isn't
provisioned.

Interestingly the charm profile will be written later on when ensuring
that the profile is there, by the provisioner API. So we don't have
to write any other code that to stop the applying of the charm
profile.

 - Add tests around provisioner task