Review noticed we were setting nextTimeout incorrectly.
It was a little bit tricky to test, because we *were* calling
Timer.Reset() with the right duration. But it lead to a situation where
we thought the timeout was later than it actually was.
TestClaimNoticesEarlyExpiry handles both cases properly. If we were
waiting a long time to timeout, when we get a claim, it will cause us to
wake up earlier, but if we then get a slightly *later* claim, it won't
push our Expiry back to the later time.