Prevent machine running state on profile error
The following commit changes how the instance poller works, because
of the fact that LXC reports a container running even if the LXD
profile applied failed.

So when applying a profile change to a container, the profile is
first added to LXD. The profile name is then added to the container
spec. Once the container is started it can sometimes error out
with an error. The problem occurs when the instance poller comes
around and asks for the instance status. LXD returns that response
as running, as it doesn't know that applying a profile that fails
should be a terminal state. We could attempt to apply the profile
again, but the LXD client doesn't tell us what the error is and
what was the cause. So retrying a profile might not be worth it,
missing device vs just unable to apply the profile at that time.

Instead we find out when the profile didn't apply and we put the
machine in to a profile error status, which can then be picked up
by the poller, which in turn can then work out if we should keep
the current error message and prevent it from become running or
we should move along.

The underlying fix for this is to get LXD to go into an ERROR state
if when applying a profile it fails, instead of a RUNNING state.