Remove HA backup/restore
The following changes removes the HA backup/restore of the
acceptance test as we no longer support it. Having talked to Rick
about this, we decided to remove this until we can re-add it back
to the CLI. This way we can then get the recovery CI tests going
green.