Fix a flaky test.
The test that we notice when a Claim has a timeout newer than our
current expiry timeout had a race. The timeout was updated after the
Claim had been processed, which meant the ensureTimeout could happen
arbitrarily far in the future.

Just changing the code to call ensureTimeout before the claim is
returned handles the race. This shouldn't affect production deployments,
because if all the claims have the same timeout, then any 'new' claim
will always be later than any other claims that we already have. And
even if we didn't, at worst it would backstop the expiry to the poll
timeout. So we might be a second behind expiring, but nothing that would
be a problem in production.