Tweak the worker/lease retry code.
It wasn't properly handling if we got 5 ErrInvalid responses in a row.
Now we require 10 invalid responses, and if we end up getting all
failures, we just deny the claim, because clearly the Primary node knows
something we don't. Ultimately, it is a failure that we are failing to
Sync with the master node.

The failure here means that the client should go into
BlockUntilLeadership call. And if we have gotten in sync by then, we
will block them. If we still haven't then they will be immediately
unblocked, and they'll try to do another claim. But it should still be
a) Correct
b) With delay backoff shouldn't introduce too much load.

This does impact other retry loops, but it shouldn't be a problem. For
Expire, we don't have ErrInvalid, as we treat that the same as if we
successfully expired it (at the very least, we know that our existing
lease is incorrect, even if it is now just owned by someone else.)