NotEmpty vs IsEmpty
Considering we only ever really want to test when a profile is not
empty (to then do something), it makes a lot more sense to flip to
an adverb as that's what we're intending to look for in most cases.

Also moved back to using profile.Empty() as we should align to how
others are using profile, so not to confuse operators.