Clean up lxdprofile core
The following commit ensures that core/lxdprofile doesn't bring any
adverse import changes into the core package. We can test this
by using TestImports in the package_test.go.

The fallout from this is that we then need to ensure we convert
the various charm, charm archive and charm info into the right
type with various shims-structs.

 - Check against nil pointers