Fix a race in remoteServer.Publish
If the connection was reset between when data was queued and the notify
channel was signalled, the worker loop wouldn't be receiving on r.data,
so publish would block indefinitely. (Naked channel sends are bad,
mmkay?) This would lock up the pubsub forwarder if a controller machine
went down, generally on the machine running the primary lease clock
updater since that's the one forwarding the most messages.

Adding a channel that gets closed when the connection is reset allows
publish to skip notifying if the connection has gone away.