Prevent calls to SetCharmProfile
So turns out SetCharmProfile isn't idempotent, so if validation
fails, then subsequent calls to SetCharmProfile will cause calls
to upgrade-charms fail.

The solution to this, is to then do the validation before hand and
ensure that we don't do any non-idempotent calls before that. This
turns out to be a bit of a pain, as we now need intermediary calls
in between.