Change SetSSHHostKeys to use buildTxn.
Rather than having it always Insert+Update, now it reads to see which
one it needs to do. The old code worked with client-side transactions,
but doesn't feel great. So switch to something that will work with
server-side transactions, and still feels a bit better. (No need to
create Txns that don't do anything.)