Fix root disk extending by polling the disk size
Creating a new VM with a non-default disk size would hang because the
after the extend disk task is successfully started, the API call to get
the task status fails with the permission error below. This means the
task waiter never returns. I haven't been able to work out the cause of
this - it seems like a bug in the VSphere API. The extend disk task
itself succeeds, we just can't request the status.

To work around this, poll the size of the VM's disk to detect when it's
been extended and return then.

(Response from the API to the PropertyCollector.WaitForUpdatesEx
call. Calls to the same method for other tasks work fine.)

&types.UpdateSet{
    DynamicData: types.DynamicData{},
    Version:     "1",
    FilterSet:   {
        {
            DynamicData: types.DynamicData{},
            Filter:      types.ManagedObjectReference{Type:"PropertyFilter", Value:"session[524ebcf8-83c1-088e-4bff-5cae1d4432f6]52cc13ca-5757-a877-cf7e-11609c8fbc26"},
            ObjectSet:   {
                {
                    DynamicData: types.DynamicData{},
                    Kind:        "enter",
                    Obj:         types.ManagedObjectReference{Type:"Task", Value:"task-153176"},
                    ChangeSet:   nil,
                    MissingSet:  {
                        {
                            DynamicData: types.DynamicData{},
                            Path:        "info",
                            Fault:       types.LocalizedMethodFault{
                                DynamicData: types.DynamicData{},
                                Fault:       &types.NoPermission{
                                    SecurityError: types.SecurityError{},
                                    Object:        types.ManagedObjectReference{Type:"Folder", Value:"group-d1"},
                                    PrivilegeId:   "System.Read",
                                },
                                LocalizedMessage: "",
                            },
                        },
                    },
                },
            },
            MissingSet: nil,
        },
    },
    Truncated: (*bool)(nil),
}