Modification Upgrade Step
The following adds a upgrade step to add missing documents for
the modification status. We jump through a few hoops here, because
we don't want to remove existing modification statuses, we only
want to add a status document, if and when a document doesn't exist.
We can work that out by checking if the machine is Alive then also
verifying that the document doesn't already exist before inserting
it.