Don't precheck instance args for machine placement
If there's a valid machine placement directive then we'll be putting
that unit onto a pre-existing machine, not starting a new cloud
instance, so we don't need to do prechecking.

This also avoids breaking the manual provider
(where PrecheckInstance always returns an error) - in that case we
always need to use machine placement because the provider can't create
new instances on demand.