Mark, sweep model cache
The following commit is just a very rough idea around marking and
sweeping values in the model cache in the case that the cache
can have old stale items in it.

The problem this fixes is when the watcher restarts, but we have
old stale content in the cache. So we should clean up anything
that hasn't been updated.