Sweep removes via dispatch
The following changes allow the remove to dispatch in the same
way that a change happens. This then allows config.Notify to work
so others can also be notified when the cache evicts entities.

It changes a few things and model isn't yet been worked out, but
that should happen in the next commit.

What's a shame is that go lacks the generics aspect so the more
things we add, the more we need to ensure that mark and sweep also
happens.