Refactor params.Error to support structured info attachments
The existing params.Error implementation hardcoded a pointer to an
"ErrorInfo" struct. Contrary to its generic-sounding name it was only
used in the case where a DischargeRequired error was emitted over an
http connection.

As params.Error can be used both by http and rpc streams it makes sense
to allow the API server to piggy-back additional information onto error
responses. One use-case for this is to implement Redirects for migrated
models; the user performs an RPC to login and the server returns back a
Redirected error together with the details of the new controller.

The params.Error changes tie together with the changes introduced to the
rpc package by the previous commits. Instead of having a hardcoded
payload, the new params.Error replaces it with a map[string]interface{}.
The original ErrorInfo struct has been renamed to DischargeRequiredErrorInfo
which better conveys its purpose and has been augmented with a AsMap
method that returns a map[string]interface{} representation that can be
attached to the Error instance.

Finally, a helper method is provided by the Error type to unmarshal the
error info into a provided struct. This approach ensures backwards
compatibility with previous client/server versions as the ErrorInfo
payload from older versions can be correctly from and to the
DischargeRequiredErrorInfo.