Unexpose entity and rename state
The following changes hide the entity type (public to private) and
along with changes around state, which is now renamed to freshness.
This should hopefully prevent confusion we have with things named
state.

Additional changes include returning an error during the sweep
phase, where a removalDelta function is missing. Note it's ok to have
a removalDelta function, but just return nil.