Adding instance mutater to machine agent
The following adds the instance mutater to the machine agent. I
don't think I've got the manifolds correct and I'm not entirely sure
what the decorators do incompleteness (halp).

The main change here is the modelmachineswatcher now takes an auth
func so that we can inject it into other locations instead of just
on the authcontroller (is this correct?)