Attach CA cert to certs observed by the api server cert watcher
Upon observing a new server cert, the watcher will decode the CA PEM cert
that was used to sign it and append it to the certificate chain list for
the server cert.

This change makes the two link certificate chain (server + CA) available
to clients connecting to the api server and allows clients to decide
whether the CA cert can be trusted when connecting to an unknown
controller.