Support untrusted CA cert verification via DialOpts callback
DialOpts has been extended with a new field called VerifyCA. If
provided, the dial code will attempt to perform a TLS handshake with the
remote server and validate its CA cert, if provided, before allowing the
connection to proceed.

If VerifyCA is provided, the dialing code will not invoke it if:
- the remote server does not present a CA cert
- the remote server presents a CA that can be verified using the system
  CA roots

If VerifyCA returns an error then the connection attempt will be aborted
before any data is exchanged with the remote server and the error will
be returned to the caller.