Add deferred cleanups to support remove-unit --force
If force has been specified, in the dyingUnit cleanup we schedule a
forceDestroyUnit to run in a minute. This gives the unit agent a chance
to perform the normal dying->dead steps and the machine agent a chance
to recall the unit and remove it. If the unit's gone by the time it
starts the forceDestroyUnit cleanup will just exit. Otherwise it
destroys all subordinates, leaves all relations, destroys all storage
attachments and marks the unit dead. (If there are subordinates or
storage that are still around then this might take multiple attempts.)

Then it schedules another cleanup to remove the unit completely (which
gives the machine agent a chance to take action if possible).