Set the delay on the force cleanup to 0 for now
Until we have the --no-wait flag plumbed through from the api server to
the state code, we can run the force cleanups immediately - this is a
better user experience. Once the flag is there we'll change it back to 1
minute.

This should still behave correctly - the unit and machine agents should
do the right thing in response to the unit or machine going away.