Handle RedirectErrors returned to the client when invoking Login via RPC
So far, there are two possible ways to receive a Redirect error in the
client state login implementation:
1) Receive an error with a CodeRedirect but no extra payload. This code
   path is currently used by the JAAS implementation and requires an
   extra RPC to the "RedirectInfo" endpoint to fetch the new controller
   information.
2) Receive an rpc.RequestError with a valid params.RedirectErrorInfo
   payload. This codepath was introduced by the previous commit and
   already includes all the required redirection info (endpoints and CA
   cert for the new controller.

In *both* cases, state.Login is expected to return an api.RedirectError
containing the necessary redirect details back to the caller. However,
for the JAAS use-case we expect the caller to *transparently connect* to
the new controller whereas for non-JAAS use-cases we want to display an
error message to the user, perhaps advising them how to access the
redirected model instead of doing this automatically for them. This
ensures that users are always aware which controller they are targeting
with their commands.

To facilitate both use-cases, the returned RedirectError is augmented
with a "FollowRedirect" attribute which is set to true for JAAS-like
redirects. This allows us to retain the current behavior of
api.NewAPIConnection for JAAS and bubble the RedirectError error up to
the cli command for non-JAAS cases.