Merge on deletion of documents
We care about deletion of the documents from the collection, so
when a Revno of -1 comes, this means that a document is being deleted
so we should send the event out.

The uniter should handle this with a not known status if that event
causes the uniter to resolve that event.