Type inference
The previous code assumed that the type passed to .One would in
fact remain the same type. This unfortunately is not the case, in
fact the type changes from applicationDoc to bson.M (alias for
map[string]interface).

This causes no end of woes, so to prevent that issue from happening
we life the query one level higher, prevent the need to pass in a
document and do the unmarshalling of the documents at the extraction
site. This also simplifies the code, so it's actually a rather good
win their as well.