Log levels for errors
Depending on certain situations we want different types of log
levels. For example an not provisioned error should be of the level
debug, as it is to be expected in the normal flow. Nothing is wrong
and the operator doesn't need to be notified at an error level.

But if applying a profile we get an error attempting to apply the
profile, we should notify the operator of the failure to do so.