Change application teardown.
Instead of having Unit and RelationUnit destroy actually cleanup the
application doc, now they queue up a Cleanup operation that will clean
up the document later. Calling application.Destroy() when it is already
Dying is no longer a no-op as it will progress the destruction further
(if all references are gone, it will remove the app completely).