Not provisioned instance mutater
The machine that we're currently watching may not be provisioned,
if that's the case we need to loop until more changes are sent.
Unfortunately we may actually end up with missing the changes that
are being sent and so the machine might end up being provisioned
incorrectly. We need to think/solve this.