Heisenbug to be solved...
It seems that running TestForceDestroyUnitDestroysSubordinates succeeds
when run in isolation (or with limited number of tests), but fails if
you run all Cleanup tests.

Main issue is part of destroyOps. If you run it while the unit may or
may not be dead, then some of the assertions can be wrong.