Run Cleanups in sorted order.
They shouldn't need to be order dependent, but most of the time in
testing they happen to run in sorted order, so the cleanups have been
written end up assuming the order.

We want to force randomization in the future, but this gets everything
passing.

Also clean up some typos in the text.