Properly handle app.Destroy() on a dead app.
The original code noticed that "oh, I'm already dying, nothing for me to
do". The new code actually goes to cleanup references, etc. However, the
stale in-memory document could still be used to call app.Destroy(), and
it needs to notice it is in an inconsistent state, and refresh itself.

There is some concern that App.Destroy() with a stale app, that refers
to a charm shared by another app, might not notice that there is nothing
that *it* should be doing. All of the documents that it wants to remove
from the database will happily remove a missing document and not fail
the transaction. And we won't notice that decref fails, because the
*other* app will still be referencing the application.

Also added more tracing, cleaned up some tests.