Fix bug(s) in raft upgrade steps
https://bugs.launchpad.net/juju/+bug/1827371

The MigrateLegacyLeases upgrade step was failing because the raft
cluster wasn't bootstrapped, which in turn was because of a bug in the
BootstrapRaft upgrade step.

That seems to have happened because:

While upgrading from 2.3 to 2.4, BootstrapRaft created the raft
directory and then failed in raft.BootstrapCluster (for unknown
reasons). The upgrade step failed but on retrying succeeded
incorrectly, because it was checking for the raft directory rather
than making sure the cluster was really bootstrapped.

That meant that 2.4 was running without a bootstrapped raft (unnoticed
because it wasn't used in 2.4). Then when upgrading to 2.5,
MigrateLegacyLeases was looking for a configuration, but that wasn't
there because of the error in the old upgrade.

There are two parts to the fix:

1. Make BootstrapRaft check for a bootstrapped cluster, not just the
directory (probably won't have any effect, since we won't be doing
another 2.4 release, but just in case).

2. Change MigrateLegacyLeases to bootstrap the cluster if it isn't
already - this will fix the problem if the 2.4 upgrade left the
cluster in a weird state.

While doing this I realised that looking for config records only in
the log store isn't robust - if there have been enough log entries they
could have been compacted into a snapshot, so we need to look there
for config and last index info as well.