Allow lxd profile test to run on aws
Some of the lxd profile tests will fail because of the assertions
with in the test on what profiles should be where. This PR fixes
this, so that we don't care what profiles land where.

It's a better fix because we use a object for instance removal,
rather than a counter, so we can visually inspect the object to
see what's left to be applied rather than just knowing something
has to be applied.