Intermittent test failure for error case
The following fixes an error case where the test would call kill
on the worker before the actual work calls kill itself. It's a race
to see whom calls kill. The substantial changes ensure that we
wait until the worker mutater context has been killed with an error
first, then we can move on.

We have had to expose some machinary, but it's not as bad as first
thought.