Nil pointer fix.
This is an easy gotcha with Go interfaces. If you have an object which
is a pointer, and you just "return var" it returns a "typed nil" which
doesn't then == nil anymore.

So the fix is to check if the pointer you have is nil, and then return a
nil interface.

Related to bug 1829014.