Change k8s storage config upgrade step to skip non-live models
We've had users with dying models that they can't remove (because the
k8s has gone away and we're trying to talk to it before removing the
model). Those models would prevent the k8s storage config step from
succeeding and block the upgrade. Skip dying models in the step.

Also drive-by fix to handle the error if the cloud config can't be
retrieved.