Bundle force
The following is an introduction of allowing the usage of force
for bundles, which then allow the overriding of series and LXD
profiles. To the extent that if there is even a validation error
when deploying a charm/bundle, the use of force can then ignore
the validation error and allow the charm/bundle be deployed.

The justification for this is that we can then reason more about
what charms and bundles do when being deployed without having to
think about the exceptions to each.