Use actions for tokens
The following uses actions on a charm, rather than using config
changes. Although the config changes ensure that a relation is set
and that the token can be read when the relation changes, the way
it sets the status and then read means that it's problematic when
reading the value in the tests. This is more evident when running
against a cloud like aws/gce, where latency comes in.

To improve this, we instead run an action for the charm to ensure
that the charm is still alive and that we can read the value from
the action result. This replicates some of the functionality of
the config setup, but the difference being that it's reliable to
work on the clouds.