Remove redunant tests and fix broken agent uninstall test
The issue with the tests was that EnsureDead() was called before the
machiner worker had a change to start. After the call to EnsureDead()
the agent will not be able to connect to the API again. As a result, the
agent would shutdown before the machiner worker started and as a result
the agent's uninstall hooks would not trigger.

To resolve the problem, I have introduced the WithAliveAgent() helper.
All it does it to start the agent, monitor its presence till the agent
becomes alive, then invoke a user defined callback and finally wait for
agent to exit (or for a timeout occur).

TestMachineAgentUninstall now uses the new helper and passes
EnsureDead() as the callback function.

The following tests have been removed by this commit:
- TestWithDeadMachine: this one is identical to
TestMachineAgentUninstall
- TestWithRemovedMachine: this one calls EnsureDead() followed by
Remove() on the machine which fails due to storage issues

One thing that I don't really like about these tests is that it takes
about 30 seconds (or 3 * coretesting.LongWait) for the agent to become
alive.
This commit removes TestWithDeadMachine