Add a timeout for lease manager stopping
Debugging logs generated when the http-server worker times out
shutting down suggest that the lease manager is failing to stop in
some cases. Add a timeout to the shutdown so we don't wait
indefinitely if a background claim or expire doesn't finish in a
timely manner.

Track how many outstanding claims and expires there are in flight and
report these in a debug log dumped on timeout, as well as in the
introspection report. Include goroutines since (if this is the only
thing stopping the httpserver from stopping) the httpserver won't dump
them when we timeout here.

The chain from httpserver to lease manager is:
* httpserver is waiting for the mux
* the mux is waiting for api-server to call ClientDone
* the api-server is waiting for a request to finish releasing a state
  back to the pool
* the release is stuck because the state in question is waiting for
  its workers to stop
* the state worker goroutine counts indicate that there's one more
  lease manager than all of the other workers (hub watcher, presence
  watcher, ping batcher)