Let the undertaker run if the environ tracker is unavailable
In that case we start with a destroyer that always fails. This means
that the undertaker can remove models that are force-destroyed, while
non-forced destructions still fail (because the cloud environment
can't be cleaned up).

Because this manifold still depends on the environ tracker, the
undertaker will be restarted if the environ tracker becomes available
later on (eg. because the credentials have been fixed).